Handle inner mutability of StatsdBackend shard mapping
This is one method of re-using the internals of a shard map and
allowing an inner mutability step. Nothing actually mutates it yet,
since I'm going to feedback here.

Upsides: loose coupling between shard updates and backend updates

Downsides: Yet another RwLock in a hot path for allowing safe swapping
of the mapping. There are other techniques we could do here like
lazy-load from another structure or an atomic load of the map, but its
somewhat harder to reason about.

Alternative: We never mutate the internals of a StatsdBackend and
instead use the outer mutability we alreday have in the `Backends`
struct to simply replace the whole thing in one shot. There is already
a RwLock here and we can still do some games with reusing the existing
clients amongst backends.