Allow re-loading config to re-build backends
This is the other method to do reloading of backend shards, by just
replacing the entire backend but transferring any existing clients
over. Old unused clients will simply be dropped and then shut down (as
we would drop the tripwire trigger).

This has more advantages than the last PR, as it allows us to update a
lot more internal state, but fixes issues with re-using clients.